The interaction between the client and the server is completed.

+++ client/model/os.go
@@ -13,25 +13,25 @@ type OSModel struct {
 	Load15         float64 `json:"load_15"`
 	MemTotal       uint64  `json:"mem_total"` // Unit: B(Byte)
 	MemUsed        uint64  `json:"mem_used"`  // Unit: B(Byte)
-	MemUsedPct     float64 `json:"mem_used_pct"`
+	MemUsedPct     uint64  `json:"mem_used_pct"`
 	SwapMemTotal   uint64  `json:"swap_mem_total"` // Unit: B(Byte)
 	SwapMemUsed    uint64  `json:"swap_mem_used"`  // Unit: B(Byte)
-	SwapMemUsedPct float64 `json:"swap_mem_used_pct"`
+	SwapMemUsedPct uint64  `json:"swap_mem_used_pct"`
 	HddTotal       uint64  `json:"hdd_total"` // Unit: B(Byte)
 	HddUsed        uint64  `json:"hdd_used"`  // Unit: B(Byte)
-	HddUsedPct     float64 `json:"hdd_used_pct"`
+	HddUsedPct     uint64  `json:"hdd_used_pct"`
 	CpuCount       int     `json:"cpu_count"`
-	CpuUsedPct     float64 `json:"cpu_used_pct"`
+	CpuUsedPct     uint64  `json:"cpu_used_pct"`
 	NetDownSpeed   uint64  `json:"net_down_speed"`  // Unit: B/s
 	NetUpSpeed     uint64  `json:"net_up_speed"`    // Unit: B/s
 	ByteRecvTotal  uint64  `json:"byte_recv_total"` // Unit: B(Byte)
 	ByteSentTotal  uint64  `json:"byte_sent_total"` // Unit: B(Byte)
-	Ping10000      uint    `json:"ping_10000"`      // Unit: ms
-	Ping10010      uint    `json:"ping_10010"`      // Unit: ms
-	Ping10086      uint    `json:"ping_10086"`      // Unit: ms
-	LostRate10000  float32 `json:"lost_rate_10000"`
-	LostRate10010  float32 `json:"lost_rate_10010"`
-	LostRate10086  float32 `json:"lost_rate_10086"`
+	Ping10000      int     `json:"ping_10000"`      // Unit: ms
+	Ping10010      int     `json:"ping_10010"`      // Unit: ms
+	Ping10086      int     `json:"ping_10086"`      // Unit: ms
+	LostRate10000  int     `json:"lost_rate_10000"`
+	LostRate10010  int     `json:"lost_rate_10010"`
+	LostRate10086  int     `json:"lost_rate_10086"`
 	Tcp            uint    `json:"tcp"`
 	Udp            uint    `json:"udp"`
 	Process        uint64  `json:"process"`
